requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
ruamel.yaml>=0.18.0
jsonschema>=4.0.0
pydantic>=2.0.0
